a lot of refactoring in the internals of MultiType renamed set<T>() to force<T>() (forces the MultiType to switch its type to T)

+++ code/branches/testing/src/libraries/core/command/Shell.cc
@@ -169,7 +169,7 @@ namespace orxonox
         else
         {
             OutputLevel level = (value ? DefaultLogLevel::Dev : DefaultLogLevel::User);
-            ModifyConfigValueExternal(this->configurableMaxLevel_, this->getConfigurableMaxLevelName(), tset, level);
+            ModifyConfigValueExternal(this->configurableMaxLevel_, this->getConfigurableMaxLevelName(), tset, static_cast<int>(level));
         }
     }
